move extract command to archiver.extract
